refactor(columnar): extract filter operator
