+++ bench/meson.build
@@ -31,6 +31,7 @@ bench_backend_src = files(
   subdir_wirelog / 'columnar/ops.c',
   subdir_wirelog / 'columnar/arithmetic.c',
   subdir_wirelog / 'columnar/expr.c',
+  subdir_wirelog / 'columnar/filter.c',
   subdir_wirelog / 'columnar/eval_stack.c',
   subdir_wirelog / 'columnar/eval.c',
   subdir_wirelog / 'columnar/arrangement.c',
