+++ tests/meson.build
@@ -817,6 +817,7 @@ backend_src = files(
   '../wirelog/columnar/ops.c',
   '../wirelog/columnar/arithmetic.c',
   '../wirelog/columnar/expr.c',
+  '../wirelog/columnar/filter.c',
   '../wirelog/columnar/eval_stack.c',
   '../wirelog/columnar/eval.c',
   '../wirelog/columnar/arrangement.c',
